+++ OneBlink.SDK/models/FormSubmissionEventConfiguration.cs
@@ -153,5 +153,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public FormSubmissionEventPdfEmailTemplate emailTemplate
+        {
+            get; set;
+        }
     }
 }