+++ OneBlink.SDK/models/FormSubmissionEventPdfEmailTemplate.cs
@@ -1,0 +1,15 @@
+using System.Collections.Generic;
+namespace OneBlink.SDK.Model
+{
+    public class FormSubmissionEventPdfEmailTemplate
+    {
+        public long id
+        {
+            get; set;
+        }
+        public List<FormSubmissionEventPdfEmailTemplateMapping> mapping
+        {
+            get; set;
+        }
+    }
+}