+++ OneBlink.SDK/models/FormSubmissionEventPdfEmailTemplateMapping.cs
@@ -1,0 +1,18 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormSubmissionEventPdfEmailTemplateMapping
+    {
+        public string type
+        {
+            get; set;
+        }
+        public string formElementId
+        {
+            get; set;
+        }
+        public string text
+        {
+            get; set;
+        }
+    }
+}